MySQL基准测试文档重构

+++ demo-mysql-n-mariadb/demo-order-management-app-sharding-jdbc/common/src/main/java/com/future/demo/dto/OrderDTO.java
@@ -6,13 +6,13 @@ import com.future.demo.entity.DeleteStatus;
 import com.future.demo.entity.Status;
 import lombok.Data;
 
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.List;
 
 @Data
 public class OrderDTO {
-    private BigInteger id;
+    /*private BigInteger id;*/
+    private Long id;
     private Long userId;
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", timezone = "GMT+8")
     private LocalDateTime createTime;
