+++ demo-mysql-n-mariadb/demo-order-management-app-sharding-jdbc/common/src/main/java/com/future/demo/entity/OrderDetailModel.java
@@ -2,12 +2,11 @@ package com.future.demo.entity;
 
 import lombok.Data;
 
-import java.math.BigInteger;
-
 @Data
 public class OrderDetailModel {
     private Long id;
-    private BigInteger orderId;
+    /*private BigInteger orderId;*/
+    private Long orderId;
     private Long userId;
     private Long productId;
     private Long merchantId;
