+++ demo-mysql-n-mariadb/demo-order-management-app-sharding-jdbc/common/src/main/java/com/future/demo/entity/OrderModel.java
@@ -2,13 +2,13 @@ package com.future.demo.entity;
 
 import lombok.Data;
 
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.List;
 
 @Data
 public class OrderModel {
-    private BigInteger id;
+    /*private BigInteger id;*/
+    private Long id;
     private Long userId;
     private LocalDateTime createTime;
     private Status status;
