+++ demo-mysql-n-mariadb/demo-order-management-app-sharding-jdbc/common/src/main/java/com/future/demo/mapper/OrderDetailMapper.java
@@ -3,7 +3,6 @@ package com.future.demo.mapper;
 import com.future.demo.entity.OrderDetailModel;
 import org.apache.ibatis.annotations.*;
 
-import java.math.BigInteger;
 import java.util.List;
 
 @Mapper
@@ -14,7 +13,8 @@ public interface OrderDetailMapper {
             "       #{item}" +
             "   </foreach>" +
             "</script>")
-    List<OrderDetailModel> list(List<BigInteger> orderIdList);
+    /*List<OrderDetailModel> list(List<BigInteger> orderIdList);*/
+    List<OrderDetailModel> list(List<Long> orderIdList);
 
     @Insert("insert ignore into t_order_detail(id,orderId,userId,productId,merchantId,amount) values(#{orderDetail.id},#{orderDetail.orderId},#{orderDetail.userId},#{orderDetail.productId},#{orderDetail.merchantId},#{orderDetail.amount})")
     int insert(@Param(value = "orderDetail") OrderDetailModel orderDetailModel);
